Delete all outdated export jobs

+++ app/Services/ExportJob.php
@@ -23,19 +23,25 @@ class ExportJob extends Base
 
     public function deleteOld(): bool
     {
-        $days = $this->getConfig()->get('exportJobsMaxDays', 21);
+        $days = $this->getConfig()->get('exportJobsMaxDays', 2);
         if ($days === 0) {
             return true;
         }
 
         // delete
-        $toDelete = $this->getEntityManager()->getRepository('ExportJob')
-            ->where(['modifiedAt<' => (new \DateTime())->modify("-$days days")->format('Y-m-d H:i:s')])
-            ->limit(0, 2000)
-            ->order('modifiedAt')
-            ->find();
-        foreach ($toDelete as $entity) {
-            $this->getEntityManager()->removeEntity($entity);
+        while (true) {
+            $toDelete = $this->getEntityManager()->getRepository('ExportJob')
+                ->where(['modifiedAt<' => (new \DateTime())->modify("-$days days")->format('Y-m-d H:i:s')])
+                ->limit(0, 2000)
+                ->order('modifiedAt')
+                ->find();
+            if (empty($toDelete[0])) {
+                break;
+            }
+
+            foreach ($toDelete as $entity) {
+                $this->getEntityManager()->removeEntity($entity);
+            }
         }
 
         // delete forever
